huangshiwei4@huawei.com

+++ interfaces/kits/native/extractortool/include/file_mapper.h
@@ -45,7 +45,7 @@ private:
     static const size_t MMAP_PROT = PROT_READ;
     static const int32_t MMAP_FLAG = MAP_SHARED;
     static const int32_t MAP_XPM = 0x40;
-    static size_t pageSize_;
+    static int32_t pageSize_;
 
     std::string fileName_;
     bool isCompressed = false;
